[TS] Leave 'a as A' as-is when it may not be null
Usually, in JS as well as TS mode, 'a as A' is converted
to a utility function call 'as(a, A)'. The ActionScript
semantics of 'as' are to return 'null' when the type check
does not succeed.
When 'a as A' is immediately de-referenced, it may not be
null, or the JS equivalent of an NPE would be raised. If
that is intended, it would be *much* better to use 'cast'
(which throws a TypeError) instead of 'as'.
"De-referenced" here means, used as the first argument in
a dot, array index or apply expression.
Thus, we assume in those cases that the 'as' expression was
only used in ActionScript to prevent a compile-time type
error, and we can use the TypeScript 'as' operator directly.
Note that this removes the run-time type check and leads
to different semantics if the NPE actually occurred and was
caught before.
Also note that only direct de-referencing usage is checked,
not a real "nullability analysis". For example, introducing
a local variable "hides" de-referencing from this heuristic:
  var foo = bar as Foo; foo.something; // not detected

Avoiding 'as' function calls not only reduces run-time
overhead, it also sometimes spares importing the class
to cast into as a value and can thus help resolve cyclic
ES5 module dependencies.