[TS] Option to use new this-usage-before-super() syntax
There is a new switch typeScriptThisBeforeSuperViaIgnore to
generate the new TypeScript syntax when in ActionScript,
'this' is used in the constructor before the super() call.
Because this is not allowed in ECMAScript, and thus not in
TypeScript, but works in Ext JS classes and cannot be avoided
in certain parts of the Ext JS programming model, we must
use some workaround.
The old workaround relied on the TypeScript compiler not
detecting usage of 'this' inside an Immediately Evaluating
Arrow Function, so the converter wrapped all such code into
such a function. However, the shortcoming of the TypeScript
compiler might be fixed some day, and the resulting syntax
was not very intuitive.
The new syntax accepts the fact that we are doing something
non-standard here and tell the TypeScript compiler by using
the directive '// @ts-expect-error', which is the stricter
sibling of '// @ts-ignore', since it fails when there is no
error, so you cannot forget to remove it together with the
next line of code it refers to.
To keep the non-standard code lines as few and as short as
possible, the converter now introduces an alias for 'this'
and uses the compiler directive only for its declaration:

    // @ts-expect-error Ext JS semantics"
    const this$ = this;

All following usages of 'this' are then substituted by
'this$', like when a function refers to its outer 'this'.