[TS] Improve support for relative [Embed] paths
The relative path used in [Embed] annotations in
ActionScript is now transformed to work in the
TypeScript output for more usage scenarios:

* resources are in a dedicated folder under
  'src', but not using the folder structure
  corresponding to the ActionScript package

* resources are outside the 'src' folder, but
  are copied into the TypeScript target folder,
  usually using the Maven <resources> element

The first scenario failed when using an
'extNamespace' to cut off in the TypeScript
path.

The second scenario just did not work, the
relative paths were kept as-is, leading to
dead references.