Rudimentary support for import directives before secondary classes
In ActionScript, "secondary" classes are declared outside
the package scope and have their own import declarations.
Unfortunately, this does not fit Jangaroo's approach to
maintain an import list per compilation unit. But we can
improve the situation to work for typical usages from
practice by at least adding "secondary" import declarations
to this global list, rather than omitting them.
In the JS output, this leads to an incorrectly placed
comment containing the import directive, but for the
rare usages of secondary classes, this should be okay.