Introduce flags for TypeScript target source format features
Instead of adding more and more top-level flags like
'typeScriptThisBeforeSuperViaIgnore' to JoocOptions
and thus as command-line options, all flags that tweak
the TypeScript output are now aggregated under the new
option 'typeScriptTargetSourceFormatFeatures'.
The value is a bit-set of the feature flags described
in the CLI help text (short) and the interface with
the corresponding constants TypeScriptTargetSourceFormatFeature
(long).

The new flags are for "simplified as-expressions" (opt-out
was missing) and "static blocks" (new).

Some integration tests have been extended to additionally
run with some flag on, expected result is then looked up
below test/resources/expected2/...