[jooc] Eventually implement coercion
Supporting coercion is especially important when
ActionScript code mixes Number and int/uint usages.
For example, a wide-spread pattern seems to be
    var u: uint = Math.random() * 10;
ActionScript takes care of converting the right-hand
side float value into an unsigned integer before
assigning it to 'u'.

Jangaroo so far treated int/uint like Number and
left the responsibility to convert floats to integers
to the developer. This leads to errors when compiling
original ActionScript code that relies on coercion.

The compiler now implements coercion for variable
assignments, including declaration initializers.
Coercion is not yet implemented for
argument-to-parameter assignment.
For TypeScript, the feature is opt-in
(TypeScriptTargetSourceFormatFeature.SAFE_COERCIONS),
so that old conversions are still reproducible with
the new converter.