Amend fix for JOO-163: Use heuristic for correct parameter sequence
https://jangaroo.myjetbrains.com/youtrack/issue/JOO-163

In Ext AS, the order of event parameters is not guaranteed to be
the declaration order of event class properties.
See comprehensive code comment in TypeScriptCodeGenerator#getEventParameters().
To test the heuristic, the example event class SomeEvent has been
augmented by another event property and renamed to follow the Ext AS
event class name pattern.