Amend fix for JOO-163: Use __PARAMETER_SEQUENCE__  if available and log error if both heuristics fail
https://jangaroo.myjetbrains.com/youtrack/issue/JOO-163

The declaration that really defines the event parameter
sequence is the value of the __PARAMETER_SEQUENCE__
constant in event classes. The heuristic is only needed
when the initializer value of the constant is not
available at compile time, because only the API stub
is used, not the source code.
Thus, we now first check for __PARAMETER_SEQUENCE__,
only if it is not available apply the heuristic, and
log a compile error if the heuristic also fails.
This is clearer than to generate code that, due to an
incorrect event parameter sequence, may fail in a subtle
way at runtime.